requirements.txt
Django==3.1.6
Pillow==8.1.0
djangorestframework==3.12.2
